requirements.txt
pymongo[srv]
discord.py
aiohttp
pillow
dnspython
asyncpg
jishaku